+++ jsonschema-maven-plugin/src/test/java/com/github/victools/jsonschema/plugin/maven/TestPackage/TestClassB.java
@@ -14,25 +14,15 @@
  * limitations under the License.
  */
 
-package com.github.victools.jsonschema.plugin.maven;
+package com.github.victools.jsonschema.plugin.maven.TestPackage;
 
-/**
- * Configuration class to hold the configuration of generator modules and their options.
- */
-public class GeneratorModule {
+import com.fasterxml.jackson.annotation.JsonClassDescription;
 
-    /**
-     * Name of the module.
-     */
-    public String name;
+@JsonClassDescription("Jackson annotation classB")
+public class TestClassB {
+    private double aDouble;
 
-    /**
-     * The fully qualified java class name of the module.
-     */
-    public String className;
-
-    /**
-     * The options to be configured on the module.
-     */
-    public String[] options;
+    public TestClassB(double aDouble) {
+        this.aDouble = aDouble;
+    }
 }